gcs-demo/function_src/requirements.txt
google-cloud-storage==1.26.0
sendgrid==6.1.3